+++ src/index.js
@@ -1,3 +1,4 @@
 export { EmulatorEngine, DEFAULT_CDN_PATH } from "./emulator-engine.js";
 export { CORE_REGISTRY, getSystemConfig, detectSystemsByExtension, systemsRequiringThreads } from "./core-registry.js";
 export { REWIND_PROFILES, getRewindProfile } from "./rewind-profiles.js";
+export { CDN_CHANNEL_OVERRIDES, getPathToData } from "./cdn-channels.js";